refactor: split code by responsibility
